Update cookie_controller.py: fix an odd "NoneType not iterable error" on rerun
The error occurs only(?) when you .get for a cookie before putting any cookies at all into the object. And, as I say, only on rerunning the code (refreshing the page

Presumably this is caused by the key being *in* the session_state, but its value being None. So, I just give a default of {} for such falsy cases.